+++ ft_strsplit.c
@@ -6,7 +6,7 @@
 /*   By: mbelalou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/24 12:45:13 by mbelalou          #+#    #+#             */
-/*   Updated: 2017/11/30 20:35:40 by mbelalou         ###   ########.fr       */
+/*   Updated: 2017/12/01 20:13:09 by mbelalou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	**ft_strsplit(const char *temp, char c)
 {
-	int		dim;
+	size_t	dim;
 	char	**resultat;
 	char	*str;
 	int		id_mot;
@@ -27,7 +27,7 @@ char	**ft_strsplit(const char *temp, char c)
 	if (resultat == NULL)
 		return (NULL);
 	id_mot = 0;
-	while (id_mot < dim)
+	while (id_mot < (int)dim)
 	{
 		resultat[id_mot] = ft_getword(str, id_mot, c);
 		id_mot++;
